test(design): update defaults and form tests
